plugins/network-device-manager/requirements.txt
paramiko>=2.11.0
netmiko>=4.1.0
cryptography>=38.0.1
pycryptodomex>=3.16.0
pyyaml>=6.0
pynetbox>=6.6.0 